:tea: test: refactor common.ts
